requirements.txt
fastapi>=0.100.0
pydantic>=2.5.1
etcd3>=0.12.0
protobuf<=3.20.3
fastapi[standard]
chromadb-client>=0.5.0
sentence-transformers>=3.0.0
scikit-learn
plotly==6.6.0
pandas==3.0.2
openai==2.30.0